refactor(446): split repository DTOs into cr-domain/src/dto.rs
Part of #436. Moved 183 lines of `*Record` / `NewVideo` / `LandmarkSummary`
/ `PoolSummary` etc. out of `cr-domain/src/repository.rs` (292 lines)
into a dedicated `cr-domain/src/dto.rs` so the trait file focuses on
ports only and drops to 112 lines.

- New `cr-domain/src/dto.rs` with all row-shaped return types.
- `repository.rs` gains `pub use crate::dto::*;` so existing import paths
  (`cr_domain::repository::RegionRecord`) keep working — downstream
  crates migrate over time without a forced churn.
- `cr-app/src/queries.rs` updated to use `cr_domain::dto::*` directly.

Remaining #446 scope (dead_code audit across ~30 sites; `build_query_string`
helper shared between films and series handlers; repository tests for the
five currently-untested repos) is deferred — architecturally orthogonal
to the DTO split and better handled in a follow-up PR that can focus on
it without mixing in unrelated concerns.
